// frontend/src/shared/api/generated/models/TeamRollupOut.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type TeamRollupOut = {
    id: number;
    teamId: number;
    date: string;
    totalSessions: number;
    avgLfi: (number | null);
    styleCounts: (Record<string, number> | null);
};